Add notation
Can we add examples // specifics about the notation of each currency. Earlier in the guides it is mentioned that some currencies are weird. It would be awesome if we give examples to reduce the weirdness

For instance:
Euro                 ``EUR``  All payment methods  2 decimals  10,00 
Japanese yen         ``JPY``  PayPal, credit card 0 decimals  10