PROC-1601 Implement Python directive to create content for example switcher
This commit creates a Python directive, called `code-block-selector`, that encapsulates all the code examples, by adding the code-block directives as body of this new directive. I also added the directive to the `get-order` documentation page so it can be tested, before applying it to all pages later on.

The directive will:
- Create the structure for the code example switcher bar, and add all supported programming languages to it.
- It will output the existing code blocks as-is, and add generic code examples for those that are missing on the documentation page. It includes a link to the respective GitHub page, if one would like to open a pull request.

Some other changes:
- Replaced the `_get-order-response` anchor link with `_create-order-response`.