Update recurring.rst
At the moment it's a bit unclear how to handle a subscription model while having to make a first-time payment. The best solution is to charge the customer the amount that will be periodically charged right away. Next, you can set up a subscription which starts after the first period (see subscription API, startDate)