Mollie Components payments no longer immediately result in a final payment state (in case 3-D Secure authentication was not available)
Instead, we now always return a `checkoutUrl` where the shopper should be redirected. It could still be the case that once redirected, 3-D Secure authentication is not available, but the result is a single flow instead of two. The merchant should always redirect to the `checkoutUrl` and error messages are available through the 'getPayment' endpoint.