Pin the version of the 'docutils' package
Version 0.18 was released today, and apparently it breaks the build. Rather than diving into the rabbit hole to see exactly what is causing it, let's pin the version of the package in order to keep the build successful. I'll create an issue for it, such that we can deal with it separately.